Rename top-petition's petition-preview to blurb
